cpuarch: move APIC constants to a common location
APIC constants are an aspect of platform architecture and should not be
embedded in guest APIC emulation code.
