+++ cpuarch/src/lib.rs
@@ -7,3 +7,4 @@ pub mod sev_status;
 pub mod snp_cpuid;
 pub mod vmsa;
 pub mod x86;
+pub mod x86apic;
