+++ kernel/src/platform/native.rs
@@ -16,7 +16,6 @@ use super::cpuid;
 use crate::address::{PhysAddr, VirtAddr};
 use crate::console::init_svsm_console;
 use crate::cpu::IrqGuard;
-use crate::cpu::apic::{ApicIcr, IcrMessageType};
 use crate::cpu::features::{Feature, cpu_get_feat, cpu_has_feat};
 use crate::cpu::irq_state::raw_irqs_disable;
 use crate::cpu::msr::write_msr;
@@ -35,6 +34,8 @@ use crate::types::PAGE_SIZE;
 #[cfg(debug_assertions)]
 use crate::types::PageSize;
 use crate::utils::MemoryRegion;
+use cpuarch::x86apic::ApicIcr;
+use cpuarch::x86apic::IcrMessageType;
 
 use bootdefs::kernel_launch::ApStartContext;
 use bootdefs::kernel_launch::SIPI_STUB_GPA;
